styling adjustments and wcag adjustments license box
