+++ packages/ndla-article-scripts/src/domHelpers.js
@@ -50,36 +50,3 @@ export const removeModifiers = (className, modifier, rootSelector = '') => {
     el.classList.remove(`${className}--${modifier}`);
   });
 };
-
-let scrollbarWidth = null;
-
-export const getScrollbarWidth = () => {
-  if (scrollbarWidth) {
-    return scrollbarWidth;
-  }
-
-  const scrollDiv = document.createElement('div');
-  scrollDiv.style.width = '100px';
-  scrollDiv.style.height = '100px';
-  scrollDiv.style.overflow = 'scroll';
-  scrollDiv.style.position = 'absolute';
-  scrollDiv.style.top = '-9999px';
-
-  document.body.appendChild(scrollDiv);
-  scrollbarWidth = scrollDiv.offsetWidth - scrollDiv.clientWidth;
-  document.body.removeChild(scrollDiv);
-
-  return scrollbarWidth;
-};
-
-export const toggleNoScroll = (enable) => {
-  const htmlElement = document.querySelector('html');
-  if (enable) {
-    const scrollWidth = getScrollbarWidth();
-    htmlElement.style.overflow = 'hidden';
-    htmlElement.style.paddingRight = `${scrollWidth}px`;
-  } else {
-    htmlElement.style.overflow = null;
-    htmlElement.style.paddingRight = null;
-  }
-}
