+++ packages/ndla-ui/src/global/objects/_objects.backdrop.scss
@@ -1,0 +1,9 @@
+.o-backdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+
+  background: rgba(1, 1, 1, 0.3);
+}
