+++ packages/ndla-ui/src/main.scss
@@ -55,6 +55,7 @@ $inuit-wrapper-width: 1024px;
 @import 'global/objects/objects.lists';
 @import 'global/objects/objects.wrapper';
 @import 'global/objects/objects.content';
+@import 'global/objects/objects.backdrop';
 
 // COMPONENTS
 // @import "modal/component.licenseToggle";
